requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cramjam==2.8.3
Django==5.1.1
django-extensions==3.2.3
django-ninja==1.3.0
dnspython==2.6.1
pydantic==2.9.1
pydantic_core==2.23.3
pymongo==4.8.0
python-snappy==0.7.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
winkerberos==0.12.0
